Migración para crear la tabla tb_detallado_simulador

+++ database/migrations/2021_09_17_082544_create_tb_detallado_simulador_table.php
@@ -27,7 +27,7 @@ class CreateTbDetalladoSimuladorTable extends Migration
             $table->integer('unidadesavender');
             $table->integer('precioventaunitario');
             $table->integer('margencontribucion');
-            $table->float('promedioponderado');
+            $table->decimal('promedioponderado',20,4);
             $table->float('puntodeequilibriototal');
             $table->float('puntodeequilibrioproducto');
             $table->foreignId('idProducto')->constrained('tb_producto');
